src/use-cases/fetch-user-check-ins-history/fetch-user-check-ins-history.ts
import { CheckIn } from "@prisma/client";

import { CheckInsRepository } from "@/repositories/check-ins-repository";

interface FetchUsersCheckInsHistoryUseCaseRequest {
  userId: string
  page: number
}

interface FetchUsersCheckInsHistoryUseCaseResponse {
  checkIns: CheckIn[]
}

export class FetchUsersCheckInsHistoryUseCase {
  constructor(private checkInsrepository: CheckInsRepository) {}

  async execute({
    userId,
    page
  }: FetchUsersCheckInsHistoryUseCaseRequest): Promise<FetchUsersCheckInsHistoryUseCaseResponse> {
    const checkIns = await this.checkInsrepository.findManyByUserId(userId, page)

    return {
      checkIns
    }
  }
}